Refactor the run_analysis
